requirements.txt
Django==3.0.8
django-environ==0.4.5
djangorestframework==3.11.0
channels==2.4.0
channels-redis==3.0.0
django-redis==4.12.1
asgiref==3.2.10
Pillow==6.1.0
